+++ file/lib/src/sentry_file.dart
@@ -53,17 +53,17 @@ class SentryFile implements File {
   }
 
   @override
-  Future<File> create({bool recursive = false}) {
+  Future<File> create({bool recursive = false, bool exclusive = false}) {
     return _wrap(
-      () async => _file.create(recursive: recursive),
+      () async => _file.create(recursive: recursive, exclusive: exclusive),
       'file.write',
     );
   }
 
   @override
-  void createSync({bool recursive = false}) {
+  void createSync({bool recursive = false, bool exclusive = false}) {
     return _wrapSync(
-      () => _file.createSync(recursive: recursive),
+      () => _file.createSync(recursive: recursive, exclusive: exclusive),
       'file.write',
     );
   }